% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_example_data.R
\name{Example_data}
\alias{Example_data}
\alias{chrZ_genome}
\alias{chrZ_gtf}
\alias{example_bams}
\title{Example Data for SpliceWiz}
\usage{
chrZ_genome()

chrZ_gtf()

example_bams(path = tempdir(), overwrite = FALSE, offline = FALSE)
}
\arguments{
\item{path}{(Default = tempdir()) The desired destination path in which to
place a copy of the files. The directory does not need to exist but its
parent directory does.}

\item{overwrite}{(Default = \code{FALSE})
Whether or not to overwrite files if they already exist
in the given path.}

\item{offline}{(Default = \code{FALSE})
Whether or not to work in offline mode. This may be suitable
if these functions have been previously run and the user wishes to run
these functions without fetching online hub resources. Default = FALSE}
}
\value{
For \code{chrZ_genome} and \code{chrZ_gtf}: returns the path to the example genome
FASTA and gene annotation GTF files

For \code{example_bams}: returns a vector specifying the location of the 6
example BAM files, copied to the given \code{path} directory. Returns NULL if
a connection to ExperimentHub could not be established, or if some BAM
files could not be downloaded.
}
\description{
A synthetic reference, with genome sequence (FASTA) and gene annotation (GTF)
files are provided, based on the genes SRSF1, SRSF2, SRSF3, TRA2A, TRA2B,
TP53 and NSUN5. These genes, with an additional 100 flanking nucleotides,
were used to construct an artificial "chromosome Z" (chrZ).
Gene annotations,
based on release-94 of Ensembl GRCh38 (hg38), were modified with
genome coordinates corresponding to this artificial chromosome.\cr\cr
Accompanying this, an example dataset was created based on 6 samples from the
Leucegene dataset (GSE67039). Raw sequencing reads were downloaded from
\href{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE67039}{GSE67039},
and were aligned to GRCh38 (Ensembl release-94) using STAR v2.7.3a. Then,
alignments belonging to the 7 genes of the chrZ genome were filtered, and the
nucleotide sequences of these alignments were realigned to the chrZ reference
using STAR.\cr\cr
}
\section{Functions}{
\itemize{
\item \code{chrZ_genome}: Returns the location of the genome.fa file of
the chrZ reference

\item \code{chrZ_gtf}: Returns the location of the transcripts.gtf
file of the chrZ reference

\item \code{example_bams}: Fetches BAMs from ExperimentHub and places
them in the given path; returns the locations of the 6 example BAM files
}}

\examples{
# returns the location of the genome.fa file of the chrZ reference

genome_path <- chrZ_genome() 

# returns the location of the transcripts.gtf file of the chrZ reference

gtf_path <- chrZ_gtf() 

# Fetches data from ExperimentHub and places them in the given path
# returns the locations of the 6 example bam files

bam_paths <- example_bams(path = tempdir())

}
