% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dash.R
\name{spliceWiz}
\alias{spliceWiz}
\title{Launches the SpliceWiz Graphics User Interface using Shiny Dashboard}
\usage{
spliceWiz(mode = c("dialog", "browser"), width = 1600, height = 900)
}
\arguments{
\item{mode}{(default \code{"dialog"}) \code{"dialog"} displays SpliceWiz in a dialog box
with specified width and height. \code{"browser"} opens SpliceWiz in a browser-
like resizable window.}

\item{width, height}{If \code{mode} is set to \code{"dialog"}, the specified width
and height of the SpliceWiz app.}
}
\value{
An interactive shinydashboard SpliceWiz app runs.
}
\description{
This function launches the SpliceWiz interactive app using Shiny Dashboard
}
\examples{
if(interactive()) {

# Launches interactive ShinyDashboard SpliceWiz app as fixed-size dialog box
    spliceWiz(mode = "dialog", width = 1600, height = 900) 

# Launches interactive ShinyDashboard SpliceWiz app as browser window
    spliceWiz(mode = "browser")

}

}
