% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covPlotly-methods.R
\name{covPlotly-class}
\alias{covPlotly-class}
\alias{getExonRanges}
\alias{getExonRanges,covPlotly-method}
\alias{setResolution}
\alias{setResolution,covPlotly-method}
\alias{showExons}
\alias{showExons,covPlotly-method}
\title{Container for plotly-based coverage plots}
\usage{
\S4method{getExonRanges}{covPlotly}(object)

\S4method{setResolution}{covPlotly}(object, resolution)

\S4method{showExons}{covPlotly}(object)
}
\arguments{
\item{object}{A covPlotly object}

\item{resolution}{How many horizontal pixels of resolution should be shown
in the final plotly object. Set to \code{0} to disable.}
}
\description{
Container for plotly-based coverage plots
}
\section{Functions}{
\itemize{
\item \code{getExonRanges(covPlotly)}: Returns a named GRanges object containing exon
ranges, without showing the associated plotly object

\item \code{setResolution(covPlotly)}: Returns a covPlotly object after setting
the output resolution of the plotly-based coverage plots.

\item \code{showExons(covPlotly)}: Returns a named GRanges object containing exon
ranges, and shows the plotly object with the annotation track showing the
named exons

}}
\seealso{
\link{plotView}
}
