% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covPlotObject-methods.R
\name{covPlotObject-class}
\alias{covPlotObject-class}
\alias{getPlotObject}
\alias{tracks,covPlotObject-method}
\alias{condition,covPlotObject-method}
\alias{plotView}
\title{Versatile coverage plots for SpliceWiz}
\usage{
getPlotObject(obj, Event, strand = c("*", "+", "-"), tracks, condition)

\S4method{tracks}{covPlotObject}(x)

\S4method{condition}{covPlotObject}(x)

plotView(
  x,
  view_start,
  view_end,
  debug = FALSE,
  oldP = covPlotly(),
  centerByEvent = FALSE,
  EventZoomFactor = 0.2,
  EventBasesFlanking = 100,
  resolution = 5000,
  trackList = list(),
  diff_stat = c("none", "t-test"),
  diffList = list(),
  reverseGenomeCoords = FALSE,
  ribbon_mode = c("sd", "sem", "ci", "none"),
  normalizeCoverage = FALSE,
  plotAnnotations = TRUE,
  plotAnnoSubTrack = TRUE,
  showExonRanges = FALSE,
  verticalLayout = c(6, 1, 1, 2),
  horizontalLayout = c(),
  filterByTranscripts = "",
  filterByEventTranscripts = FALSE,
  filterByExpressedTranscripts = TRUE,
  condenseTranscripts = FALSE,
  plotJunctions = TRUE,
  plotJuncPSI = FALSE,
  junctionThreshold = 0.01,
  plotRanges = GRanges(),
  rangesBasesFlanking = 100,
  usePlotly = FALSE,
  ...
)
}
\description{
Here, we implement fast and versatile ggplot and plotly based
coverage and sashimi plots. Users can plot with unlimited number
of individual, individual-normalized, or group-normalized tracks.
Also implemented is user-defined group-comparison differential plots
(including t-test plots). Additionally, users can generate ggplots
subsetted by exon groups. See details below.
}
\details{
The typical pipeline for plotting versatile coverage plots is as follows:
\itemize{
\item A \code{covDataObject} is generated by calling \code{getCoverageData()} using an
input \code{NxtSE} object. This step retrieves coverage, junction counts and
normalization data for the relevant genomic region being queried
\item A \code{covPlotObject} is generated by calling \code{getPlotObject()} using an
input \code{covDataObject}. This step retrieves alternative splicing event
specific data, such as normalized coverages, or group combined coverages.
\item Plots can be generated by calling \code{plotView()} using a \code{covPlotObject}.
Interactive plotly plots can be generated by setting \code{usePlotly = TRUE},
otherwise, static plots are generated. For interactive plots, a \code{covPlotly}
object is returned, which contains raw data which is downsampled by
pixel resolution prior to plotting for performance reasons.
}

Tracks are now versatile (unlimited). Samples are retrieved by individual
sample names at \code{getCoverageData()}. If \code{condition} is set in
\code{getPlotObject()}, track names are defined by their condition categorical
names; otherwise, tracks are named by individual samples when retrieved using
\code{getPlotObject()}.
\itemize{
\item When calling \code{plotView()}, \code{trackList} by default displays all tracks as
ordered in the \code{covPlotObject}. Users can supply a vector containing either
the track names (or numbers, as ordered in the \code{covPlotObject}).
Alternatively, multiple traces can be stacked in a single track by
using a list, e.g. \code{trackList = list(A = c(1,2,3), B = c(4,5,6))}.
\item For differential comparisons, \code{diffList} takes a list of pairs of samples.
For example, if \code{trackList = list("A", "B")}, then setting
\code{diffList = list(c(1, 2))} will compare groups "A" and "B". This is only
activated by setting \code{diff_stat} to anything other than \code{none}. For now,
only \code{t-test} is supported.
}

\code{plotView()} supports plotting by exon ranges, for which only static plots
are currently supported. The workflow for generating such a plot is as
follows:
\itemize{
\item A GRanges object is returned by the \code{plotView()} function and setting
\code{showExonRanges = TRUE}. \code{plotView()} will simultaneously show an
annotation plot of exons labelled by their "exon names", which is the
transcript name appended with "-E" followed by the exon number.
\item If \code{plotView()} is called and \code{usePlotly = TRUE} is set, a \code{covPlotly}
object is returned. Calling \code{showExons()} on this object will display
a plotly plot showing exon names, and returning a GRanges object of exon
ranges.
\item Exon ranges can be supplied to the \code{plotView()} function  by setting
the \code{plotRanges} parameter as a GRanges object. This will generate a static
plot showing coverage plots segmented by exons.
}
}
\section{Functions}{
\itemize{
\item \code{getPlotObject()}: Generates a covPlotObject object from a
covDataObject. Allows users to change parameters such as viewing window,
conditions, tracks, and other parameters, for customizing plot parameters

\item \code{tracks(covPlotObject)}: Returns the tracks specified in the
covPlotObject object

\item \code{condition(covPlotObject)}: Returns the condition specified in the
covPlotObject object

\item \code{plotView()}: Creates a coverage plot using the stored
data in the covPlotObject

}}
\examples{
se <- SpliceWiz_example_NxtSE(novelSplicing = TRUE)

# Assign annotation of the experimental conditions
colData(se)$treatment <- rep(c("A", "B"), each = 3)

# Retrieve coverage data for all samples for the gene "SRSF3" (and surrounds)

dataObj <- getCoverageData(
    se,
    Gene = "SRSF3",
    tracks = colnames(se)
)

# Retrieves raw / normalized coverage / junction data for the 
# specified SRSF3 skipped exon event:

plotObj_samples <- getPlotObject(
    dataObj,
    Event = "SE:SRSF3-203-exon4;SRSF3-202-int3"
)

# Retrieves data for samples grouped by the specified condition

plotObj_group <- getPlotObject(
    dataObj,
    Event = "SE:SRSF3-203-exon4;SRSF3-202-int3",
    condition = "treatment",
    tracks = c("A", "B")
)

# Show static ggplots

plotView(plotObj_samples)

plotView(plotObj_group, centerByEvent = TRUE)

# Plot junctions using PSI estimates based on individual junction SpliceOver
# metrics

plotView(plotObj_group, centerByEvent = TRUE, plotJuncPSI = TRUE)


# Show normalized coverages, individual samples stacked in grouped tracks

plotView(
    plotObj_samples, 
    normalizeCoverage = TRUE, 
    trackList = list(A = c(1,2,3), B = c(4,5,6))
)

# Show stacked group comparisons with t-test

plotView(
    plotObj_group, 
    trackList = list(c(1,2)),
    diffList = list(c("A", "B")),
    diff_stat = "t-test"
)

# Show interactive plotly:

if(interactive()) {
    p <- plotView(plotObj_samples, usePlotly = TRUE)
    show(p)
}

# Show exons with coverage plot

# static:
gr <- plotView(plotObj_samples, showExonRanges = TRUE)

# interactive:
if(interactive()) {
    p <- plotView(plotObj_samples, usePlotly = TRUE)
    gr <- showExons(p)
}

# Plot coverage by exons

p <- plotView(plotObj_samples, 
    plotRanges = gr[c("SRSF3-203-E3", "SRSF3-203-E4", "SRSF3-203-E5")],
    horizontalLayout = c(1,1,1)
)

}
\seealso{
\link{getCoverageData}
}
