% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covPlotObject-methods.R
\name{covPlotObject-class}
\alias{covPlotObject-class}
\alias{getCovPlotObject}
\alias{plotCPO,covPlotObject-method}
\title{Container to hold ready-to-plot data for coverage plots}
\usage{
getCovPlotObject(
  obj,
  Event,
  view_start,
  view_end,
  strand = c("*", "+", "-"),
  tracks,
  condition,
  ribbon_mode = c("sd", "ci", "sem", "none"),
  diff_mode = c("none", "t-test"),
  reverseGenomeCoords = FALSE,
  condensed = FALSE,
  selected_transcripts = "",
  plot_key_isoforms = FALSE
)

\S4method{plotCPO}{covPlotObject}(
  x,
  plotRanges = NULL,
  tracks = NULL,
  plotJunctions = TRUE,
  plotAnnotations = TRUE,
  plotAnnoSubTrack = TRUE,
  stackTracks = FALSE,
  plotDiffTrack = TRUE,
  interactive = FALSE,
  resolution = 10000,
  junctionThreshold = 0.01,
  ...
)
}
\description{
This object is generated by calling ? on covDataObjects.
}
\section{Functions}{
\itemize{
\item \code{getCovPlotObject()}: Generates a covPlotObject object from a
covDataObject. Allows users to change parameters such as viewing window,
conditions, tracks, and other parameters, for customizing plot parameters

\item \code{plotCPO(covPlotObject)}: Creates a coverage plot using the stored
data in the covPlotObject

}}
\examples{
se <- SpliceWiz_example_NxtSE(novelSplicing = TRUE)

# Assign annotation of the experimental conditions
colData(se)$treatment <- rep(c("A", "B"), each = 3)

dataObj <- getCoverageData(
    se,
    Event = "SE:SRSF3-203-exon4;SRSF3-202-int3",
    tracks = colnames(se)
)

# covPlotObject uses same parameters as covDataObject, unless arguments
# overridden by parameters in getCovPlotObject
plotObj_samples <- getCovPlotObject(
    dataObj
)

# Generate group plot
plotObj_group <- getCovPlotObject(
    dataObj,
    condition = "treatment",
    tracks = c("A", "B")
)

}
\seealso{
\link{getCoverageData}
}
