% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covPlotObject-methods.R
\name{covPlotObject-class}
\alias{covPlotObject-class}
\alias{getPlotObject}
\alias{tracks,covPlotObject-method}
\alias{condition,covPlotObject-method}
\alias{plotView}
\title{Container to hold ready-to-plot data for coverage plots}
\usage{
getPlotObject(obj, Event, strand = c("*", "+", "-"), tracks, condition)

\S4method{tracks}{covPlotObject}(x)

\S4method{condition}{covPlotObject}(x)

plotView(
  x,
  view_start,
  view_end,
  debug = TRUE,
  centerByEvent = FALSE,
  EventZoomFactor = 0.2,
  EventBasesFlanking = 100,
  trackList = list(),
  diff_stat = c("none", "t-test"),
  diffList = list(),
  reverseGenomeCoords = FALSE,
  ribbon_mode = c("sd", "sem", "ci", "none"),
  normalizeCoverage = FALSE,
  plotAnnotations = TRUE,
  plotAnnoSubTrack = TRUE,
  verticalLayout = c(6, 1, 1, 2),
  horizontalLayout = c(),
  filterByTranscripts = "",
  filterByEventTranscripts = FALSE,
  filterByExpressedTranscripts = TRUE,
  condenseTranscripts = FALSE,
  plotJunctions = TRUE,
  junctionThreshold = 0.01,
  plotRanges = GRanges(),
  interactive = FALSE,
  ...
)
}
\description{
This object is generated by calling ? on covDataObjects.
}
\section{Functions}{
\itemize{
\item \code{getPlotObject()}: Generates a covPlotObject object from a
covDataObject. Allows users to change parameters such as viewing window,
conditions, tracks, and other parameters, for customizing plot parameters

\item \code{tracks(covPlotObject)}: Returns the tracks specified in the
covPlotObject object

\item \code{condition(covPlotObject)}: Returns the condition specified in the
covPlotObject object

\item \code{plotView()}: Creates a coverage plot using the stored
data in the covPlotObject

}}
\examples{
se <- SpliceWiz_example_NxtSE(novelSplicing = TRUE)

# Assign annotation of the experimental conditions
colData(se)$treatment <- rep(c("A", "B"), each = 3)

dataObj <- getCoverageData(
    se,
    Gene = "SRSF3",
    tracks = colnames(se)
)

# covPlotObject uses same parameters as covDataObject, unless arguments
# overridden by parameters in getPlotObject
plotObj_samples <- getPlotObject(
    dataObj,
    Event = "SE:SRSF3-203-exon4;SRSF3-202-int3"
)

plotView(plotObj_samples)

# Generate group plot
plotObj_group <- getPlotObject(
    dataObj,
    Event = "SE:SRSF3-203-exon4;SRSF3-202-int3",
    condition = "treatment",
    tracks = c("A", "B")
)

plotView(plotObj_group, centerByEvent = TRUE)

}
\seealso{
\link{getCoverageData}
}
