% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{setSWthreads}
\alias{setSWthreads}
\title{Sets the number of threads used by SpliceWiz}
\usage{
setSWthreads(threads = 0)
}
\arguments{
\item{threads}{(default \code{0}) The number of threads for SpliceWiz to use.
Set as \code{0} to use the recommended number of threads appropriate for the
system (approximately half the available threads)}
}
\value{
Nothing.
}
\description{
SpliceWiz uses the computationally efficient packages \code{fst} and \code{data.table}
to compute file and data operations, respectively. Both packages make use
of parallelisation. If excessive number of threads are allocated, it may
impact the running of other operations on your system. Use this function to
manually allocate the desired number of threads
}
\examples{
setSWthreads(0)
}
